TotalPriceをコンポーネントに切り出す(#31)
